+++ apps/ello_v3/lib/ello_v3/schema/network_types.ex
@@ -1,0 +1,10 @@
+defmodule Ello.V3.Schema.NetworkTypes do
+  use Absinthe.Schema.Notation
+
+  object :user do
+    field :id, :id
+    field :username, :string
+    field :name, :string
+  end
+end
+
